+++ apps/stripe/.env.example
@@ -1,3 +1,6 @@
+# See src/lib/env.ts for exhaustive list of supported and required env variables
+
+
 # The key used for metadata encryption.
 SECRET_KEY=
