+++ apps/stripe/src/app/api/stripe/webhook/route.ts
@@ -15,7 +15,7 @@ import { saleorApp } from "@/lib/saleor-app";
 import { appConfigRepoImpl } from "@/modules/app-config/repositories/app-config-repo-impl";
 import { TransactionEventReporter } from "@/modules/saleor/transaction-event-reporter";
 import { StripeWebhookSignatureValidator } from "@/modules/stripe/stripe-webhook-signature-validator";
-import { transactionRecorder } from "@/modules/transactions-recording/transaction-recorder-impl";
+import { transactionRecorder } from "@/modules/transactions-recording/repositories/transaction-recorder-impl";
 
 const useCase = new StripeWebhookUseCase({
   appConfigRepo: appConfigRepoImpl,
