+++ apps/stripe/src/app/api/stripe/webhook/stripe-event-handlers/payment-intent-succeeded-handler.ts
@@ -13,7 +13,7 @@ import {
   StripePaymentIntentStatusValidationError,
 } from "@/modules/stripe/stripe-payment-intent-status";
 import { mapPaymentIntentStatusToTransactionResult } from "@/modules/transaction-result/map-payment-intent-status-to-transaction-result";
-import { RecordedTransaction } from "@/modules/transactions-recording/transaction-recorder";
+import { RecordedTransaction } from "@/modules/transactions-recording/recorded-transaction";
 
 type PossibleErrors = InstanceType<
   | typeof SaleorMoney.ValidationError
