+++ apps/stripe/src/modules/transactions-recording/recorded-transaction.ts
@@ -1,12 +1,9 @@
-import { Result } from "neverthrow";
-
-import { BaseError } from "@/lib/errors";
-import { ResolvedTransationFlow } from "@/modules/resolved-transaction-flow";
+import { SaleorTransationId } from "@/modules/saleor/saleor-transaction-id";
 import { PaymentMethod } from "@/modules/stripe/payment-methods/types";
 import { StripePaymentIntentId } from "@/modules/stripe/stripe-payment-intent-id";
 
+import { ResolvedTransationFlow } from "../resolved-transaction-flow";
 import { SaleorTransationFlow } from "../saleor/saleor-transaction-flow";
-import { SaleorTransationId } from "../saleor/saleor-transaction-id";
 
 /**
  * Holds transaction that app records during it's lifetime.
@@ -35,36 +32,4 @@ export class RecordedTransaction {
     this.resolvedTransactionFlow = args.resolvedTransactionFlow;
     this.selectedPaymentMethod = args.selectedPaymentMethod;
   }
-}
-
-export const TransactionRecorderError = {
-  PersistenceNotAvailable: BaseError.subclass("TransactionRecorder.PersistenceNotAvailableError", {
-    props: {
-      _internalName: "TransactionRecorder.PersistenceNotAvailableError",
-    },
-  }),
-  /**
-   * Current assumption is that transaction MUST exist before reading it.
-   * If not, it's something wrong with the business logic
-   */
-  TransactionMissingError: BaseError.subclass("TransactionRecorder.TransactionMissingError", {
-    props: {
-      _internalName: "TransactionRecorder.TransactionMissingError",
-    },
-  }),
-};
-
-export type TransactionRecorderError = InstanceType<
-  | typeof TransactionRecorderError.PersistenceNotAvailable
-  | typeof TransactionRecorderError.TransactionMissingError
->;
-
-export interface TransactionRecorder {
-  recordTransaction(
-    transaction: RecordedTransaction,
-  ): Promise<Result<null, TransactionRecorderError>>;
-
-  getTransactionByStripePaymentIntentId(
-    id: StripePaymentIntentId,
-  ): Promise<Result<RecordedTransaction, TransactionRecorderError>>;
 }
